Moved methods back into POM
