+++ test/e2e/page-objects/pages/swap/swap-page.ts
@@ -225,6 +225,27 @@ class SwapPage {
     );
   }
 
+  async verifyQuote(options: {
+    swapFrom: string;
+    swapTo: string;
+    amount: number;
+  }): Promise<void> {
+    await this.checkQuoteIsDisplayed();
+    await this.checkSourceToken(options.swapFrom);
+    await this.checkDestinationToken(options.swapTo);
+
+    const swapFromAmount = await this.getFromAmountValue();
+    assert.equal(swapFromAmount, options.amount.toString());
+
+    const swapToAmount = await this.getToAmountValue();
+    const normalizedSwapToAmount = Number(swapToAmount.replace(/,/gu, ''));
+    assert.equal(
+      normalizedSwapToAmount > 0,
+      true,
+      `Expected destination amount to be > 0 but got ${swapToAmount}`,
+    );
+  }
+
   async checkQuoteIsDisplayedWithoutNetworkFee(options?: {
     timeout?: number;
   }): Promise<void> {
