+++ test/e2e/tests/swaps/swaps-notifications.spec.ts
@@ -13,7 +13,7 @@ import {
   BRIDGE_FEATURE_FLAGS_WITH_SSE_ENABLED,
   DEFAULT_BRIDGE_FEATURE_FLAGS,
 } from '../bridge/constants';
-import { checkNotification } from '../../page-objects/flows/swap.flow';
+import SwapPage from '../../page-objects/pages/swap/swap-page';
 
 const MUSD_MALICIOUS_SECURITY_DATA = {
   type: 'Malicious',
@@ -148,10 +148,11 @@ describe('Swaps - notifications', function () {
 
         await bridgeQuotePage.setCustomSlippage('0.1');
 
-        await checkNotification(driver, {
-          title: 'Low slippage',
-          text: 'A value this low (0.1%) may result in a failed swap',
-        });
+        const swapPage = new SwapPage(driver);
+        await swapPage.checkNotificationBanner(
+          'Low slippage',
+          'A value this low (0.1%) may result in a failed swap',
+        );
       },
     );
   });
